Fix warning messages bug (issue#35)
Remove isTRUE clause and move the condition statement to chart.TimeSeries
which is always called by other chart.* functions. So the warning message
will only be triggered as users specify the deprecated argument(s) in
those functions. Besides, transparency of the shading area in
chart.TimeSeries.base is set to 0.3 to avoid blocking the series chart.