refactor function to use SharpeRatio.annualized
after report of odd (x/100) *100 scaling in original formulation.  This scaling is not present in Bacon.  Larger refactor was performed because tghere was quite a bit of replication from the SharpeRatio functions, and SharpeRatio.annualized should be more robust to differeing input data and frequencies.

Credit and thanks for report to Thomas Raffinot traffinot <at> gmail <dot> com