Make CAPM.beta.bull/bear return NA if all Rb are -/+
The main goal of this fix is to prevent failure of the CAPM.table
function when it's individual parts have errors. The pull request
implements two minor patches to fix CAPM.beta.bear and CAPM.beta.bull
so the functions return NA and a warning message (rather than an error)
when no positive or negative values, respectively, are present in the
Rb subset within the linear model within the .beta function. As a
result, the user that implements CAPM.table will get the desired table
output with an NA in the results if CAPM.beta.bull or CAPM.beta.bear
functions fail.

Fixes #67.