- simplify Up/Dowen capture code
Bacon defines up capture and down capture as the ratio of means.  We use the ratio of sums, which is more computationally efficient, and is an equivalent calculation, factoring our the number of opbservations used as the denominator to the mean returns.

Prior code called these cumRa and cumRb, which was confusing, and caused an additional unecessary assignment.  I've removed that for clarity and efficiency.