Fix weight_dates if index is POSIXct
The weight_dates is incorrect if 'R' is an xts object with a POSIXct
index and 'rebalance_on' is specified, because calling c() on a Date
vector and a POSIXct vector produces weird results:

R> c(Sys.Date(), Sys.time())
[1] "2017-06-04"    "4099558-09-23"

The fix is to explicitly call as.Date() on the POSIXct vector before
calling c().

Fixes #72.