Ensure 'Suggests' packages are available
There were several places we checked whether a package was on the
search path and added it to the search path if it wasn't.  This is now
discouraged in place of only loading the namespace and accessing the
necessary functions explicitly with `::`.

Uncomment the check for Hmisc in Omega.R, since Hmisc::Ecdf.default()
is called just a few lines later.