Do not use (the unexported) current.xts_chob()
This function is not exported, and should not be accessed by users.
Packages outside of xts should use the public plotting API.

We can do this in chart.Drawdown() by evaluating its colorset argument
to get its value. It doesn't appear chart.BarVar() needs the xts
plotting object anyway, so just remove the calls.

Also re-arrange the order of the multiple if-statements so similar
function calls are closer together.