initialize drawdownpeaks vector to NA
The only functional change in this commit is an initialization of the drawdownpeaks vector to NA to avoid a warning on setting a length to a NULL vector.

This function was also a complete disaster of indentation, so I cleaned that up too for my future sanity. Sorry commit history.