Integration of Standard Errors for PerformanceAnalytics (#137)
This is a major update to PerformanceAnalytics: the user can now compute standard errors to see how accurate the risk and performances estimators (e.g., Sharpe Ratio, Expected Shortfall) are. The integration is done via the R package RPESE (https://cran.r-project.org/web/packages/RPESE/index.html), which is suggested by PerformanceAnalytics. If RPESE is installed, the user can refer to the relevant vignette to see how to use this new functionality.

* Add capability to compute and return standard errors. Integration of RPESE package into PA. Update documentation of functions.

* Add new robust cleaning option based on RobStatTM package, and updated documentation accordingly.

* Make SemiSD visible to users.

* Add vignette for standard errors computation.

* Fix vignette for standard errors.

* Fix standard errors vignette.

* Fix documentation and fix default for SE argument.

* Update vignette files.

* Remove warnings. Enforce arguments to match those if there is SE computation.

* Fix vignette for PA standard errors.

* Fix bug with robust filtering option.

* Change position for control settings.

* Fix all the devtools::check() issues. Also change name from mean.arithmetic to Mean.arithmetic.

* Change name for RPESE.control function. Update vignette for changes.