update roxygen docs
The biggest change here is the migration of
PerformanceAnalytics-package to roxygen comments in the PerformanceAnalytics-package.R file.

This is complicated by the previously known limitation of Rd2roxygen that it does not parse multiple \section{}{} constructs from .Rd files.  So all these sections need to be migrated and formatted by hand.

Now that this is done, it should be easier to maintain, and perhaps even to incorporate better math and graphics in more modern Rd files.

Other Rd files were updated as well, these should only be formatting changes from a newer roxygen2 version.