2 Bug Fixes: Remove fit.models warning and Add correct chart.SFM
Removed warnings from fit.models using a warn=TRUE parameter in fm.add.class function.
The chart.SFM with lm.outliers=T gave wrong graph, i.e. the outliers weren't correctly marked.
chart.SFM with robust outliers had a problem that it was using the final robust scale of the residuals, but what was needed was the robust
scale of the initial S-estimator for lmrobdetMM(), which is called scale.S in the fitted model object, which is
used for fitting and outlier rejection.